+++ sdk/search/azure-search-documents/azure/search/documents/indexes/_search_indexer_client.py
@@ -111,6 +111,8 @@ class SearchIndexerClient(HeadersMixin):  # pylint: disable=R0904
         indexer: SearchIndexer,
         *,
         match_condition: MatchConditions = MatchConditions.Unconditionally,
+        skip_indexer_reset_requirement_for_cache: Optional[bool] = None,
+        disable_cache_reprocessing_change_detection: Optional[bool] = None,
         **kwargs: Any
     ) -> SearchIndexer:
         """Creates a new indexer or updates an indexer if it already exists.
@@ -132,7 +134,13 @@ class SearchIndexerClient(HeadersMixin):  # pylint: disable=R0904
         kwargs.update(access_condition)
         name = indexer.name
         result = self._client.indexers.create_or_update(
-            indexer_name=name, indexer=indexer, prefer="return=representation", error_map=error_map, **kwargs
+            indexer_name=name,
+            indexer=indexer,
+            prefer="return=representation",
+            error_map=error_map,
+            skip_indexer_reset_requirement_for_cache=skip_indexer_reset_requirement_for_cache,
+            disable_cache_reprocessing_change_detection=disable_cache_reprocessing_change_detection,
+            **kwargs
         )
         return result
 
@@ -281,7 +289,12 @@ class SearchIndexerClient(HeadersMixin):  # pylint: disable=R0904
 
     @distributed_trace
     def reset_documents(
-        self, indexer: Union[str, SearchIndexer], keys_or_ids: DocumentKeysOrIds, **kwargs: Any
+        self,
+        indexer: Union[str, SearchIndexer],
+        keys_or_ids: DocumentKeysOrIds,
+        *,
+        overwrite: bool = False,
+        **kwargs: Any
     ) -> None:
         """Resets specific documents in the datasource to be selectively re-ingested by the indexer.
 
@@ -302,7 +315,7 @@ class SearchIndexerClient(HeadersMixin):  # pylint: disable=R0904
             name = indexer.name  # type: ignore
         except AttributeError:
             name = indexer
-        return self._client.indexers.reset_docs(name, **kwargs)
+        return self._client.indexers.reset_docs(name, overwrite=overwrite, **kwargs)
 
     @distributed_trace
     def get_indexer_status(self, name: str, **kwargs: Any) -> SearchIndexerStatus:
@@ -358,6 +371,7 @@ class SearchIndexerClient(HeadersMixin):  # pylint: disable=R0904
         data_source_connection: SearchIndexerDataSourceConnection,
         *,
         match_condition: MatchConditions = MatchConditions.Unconditionally,
+        skip_indexer_reset_requirement_for_cache: Optional[bool] = None,
         **kwargs: Any
     ) -> SearchIndexerDataSourceConnection:
         """Creates a new data source connection or updates a data source connection if it already exists.
@@ -381,6 +395,7 @@ class SearchIndexerClient(HeadersMixin):  # pylint: disable=R0904
             data_source=packed_data_source,
             prefer="return=representation",
             error_map=error_map,
+            skip_indexer_reset_requirement_for_cache=skip_indexer_reset_requirement_for_cache,
             **kwargs
         )
         # pylint:disable=protected-access
@@ -584,6 +599,8 @@ class SearchIndexerClient(HeadersMixin):  # pylint: disable=R0904
         skillset: SearchIndexerSkillset,
         *,
         match_condition: MatchConditions = MatchConditions.Unconditionally,
+        skip_indexer_reset_requirement_for_cache: Optional[bool] = None,
+        disable_cache_reprocessing_change_detection: Optional[bool] = None,
         **kwargs: Any
     ) -> SearchIndexerSkillset:
         # pylint:disable=protected-access
@@ -614,6 +631,8 @@ class SearchIndexerClient(HeadersMixin):  # pylint: disable=R0904
             skillset=skillset_gen,
             prefer="return=representation",
             error_map=error_map,
+            skip_indexer_reset_requirement_for_cache=skip_indexer_reset_requirement_for_cache,
+            disable_cache_reprocessing_change_detection=disable_cache_reprocessing_change_detection,
             **kwargs
         )
         return SearchIndexerSkillset._from_generated(result)  # pylint:disable=protected-access