+++ sdk/ai/azure-ai-resources/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 1.0.0b6 (Unreleased)
+## 1.0.0b8 (Unreleased)
 
 ### Features Added
 
@@ -9,6 +9,18 @@
 ### Bugs Fixed
 
 ### Other Changes
+
+## 1.0.0b7 (2024-02-07)
+
+### Other Changes
+
+- Bug fixes
+
+## 1.0.0b6 (2024-02-06)
+
+### Other Changes
+
+- Bug fixes
 
 ## 1.0.0b5 (2024-02-01)